[ui] #95 가이드 이미지에 그라데이션 적용
