+++ app/src/main/res/drawable/shape_white_gradient_rect.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <gradient
+        android:startColor="@color/white"
+        android:endColor="@android:color/transparent"
+        android:angle="90" />
+</shape>
