+++ app/src/main/res/layout/fragment_winey_feed_guide.xml
@@ -67,14 +67,38 @@
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@id/tv_guide_title" />
 
-        <ImageView
+        <FrameLayout
             android:layout_width="0dp"
             android:layout_height="0dp"
             android:layout_marginTop="23dp"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/tv_guide_detail" />
+            app:layout_constraintTop_toBottomOf="@id/tv_guide_detail">
+
+            <ImageView
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                android:src="@drawable/img_guide_wineyfeed" />
+
+            <ImageView
+                android:layout_width="match_parent"
+                android:layout_height="215dp"
+                android:layout_gravity="bottom"
+                android:src="@drawable/shape_white_gradient_rect" />
+
+            <androidx.appcompat.widget.AppCompatButton
+                android:id="@+id/btn_guide_next"
+                android:layout_width="match_parent"
+                android:layout_height="46dp"
+                android:layout_gravity="bottom"
+                android:layout_marginHorizontal="16dp"
+                android:layout_marginBottom="16dp"
+                android:background="@drawable/shape_purple_fill_10_rect"
+                android:stateListAnimator="@null"
+                android:text="@string/guide_next_btn_text"
+                android:textColor="@color/white" />
+        </FrameLayout>
 
     </androidx.constraintlayout.widget.ConstraintLayout>
